% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business.R
\docType{data}
\name{business}
\alias{business}
\title{Benin Business Environment and Labour related Statisticsm}
\format{
An object of class \code{data.frame} with 27 rows and 132 columns.
}
\source{
\href{https://knoema.com/JOBS2018OCT/business-environment-and-labour-related-statistics?regionId=BJ&accesskey=lvqipff}{Business}
}
\usage{
data(business)
}
\description{
Data cited at: The World Bank https://datacatalog.worldbank.org/
Topic: Jobs Publication: https://datacatalog.worldbank.org/dataset/jobs
License: http://creativecommons.org/licenses/by/4.0/
The World Bank Jobs Statistics Over 150 indicators on labor-related topics,
covering over 200 economies from 1990 to present.
}
\examples{
data(business)
}
\references{
World Bank (2020) January 2020
(\href{https://data.worldbank.org/data-catalog/jobs}{World Bank})
}
\keyword{datasets}
