% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climat2.R
\docType{data}
\name{climat2}
\alias{climat2}
\title{Climatologie au Benin}
\format{An object of class \code{data.frame} with 204 rows and 32 columns.}
\source{
\href{https://benin.opendataforafrica.org/data/#topic=Climatologie}{Climatologie}
}
\usage{
data(climat2)
}
\description{
Climatologie au Benin
}
\examples{
data(climat2)
}
\references{
Benin Portail des donnees
Derniere publication 2017-2018
}
\keyword{datasets}
