% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_violences.R
\docType{data}
\name{genre_violences}
\alias{genre_violences}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{data.frame} with 8 rows and 24 columns.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(genre_violences)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(genre_violences)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
