% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\docType{data}
\name{mortality_rate}
\alias{mortality_rate}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 13 columns.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(mortality_rate)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(mortality_rate)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
