% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_by_age.R
\docType{data}
\name{pop_by_age}
\alias{pop_by_age}
\title{Population by Single Year Age Groups - Custom Region - Benin}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3640 rows and 9 columns.}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(pop_by_age)
}
\description{
Population by Single Year Age Groups - Custom Region - Benin
}
\examples{
data(pop_by_age)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.census.gov/data-tools/demo/idb/region.php?T=15&RT=0&A=separate&Y=2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018&C=BN&R=0}{PubMed})
}
\keyword{datasets}
