% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climat.R
\docType{data}
\name{climat}
\alias{climat}
\title{Climatologie au Benin}
\format{An object of class \code{data.frame} with 7 rows and 55 columns.}
\source{
\href{https://benin.opendataforafrica.org/data/#topic=Climatologie}{Climatologie}
}
\usage{
data(climat)
}
\description{
Climatologie au Benin
}
\examples{
data(climat)
}
\references{
Benin Portail des donnees
Derniere publication 2017-2018
}
\keyword{datasets}
