% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education}
\alias{education}
\title{Benin Education Statistics}
\format{
An object of class \code{data.frame} with 64 rows and 2081 columns.
}
\source{
\href{https://knoema.com/WBEDS2017Jun/education-statistics?regionId=BJ&accesskey=crhlebc}{Education Statistics}
}
\usage{
data(education)
}
\description{
Benin Education Statistics
}
\examples{
data(education)
}
\references{
World Bank (2019) December 2019
(\href{https://knoema.com/atlas/sources/WB}{World Bank})
}
\keyword{datasets}
