% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\docType{data}
\name{infra}
\alias{infra}
\title{National Infrastructure Database}
\format{
An object of class \code{data.frame} with 15 rows and 744 columns.
}
\source{
\href{https://knoema.com/NTLDT2016/national-infrastructure-database?regionId=BJ&accesskey=nmjmgjf}{Infrastructure}
}
\usage{
data(infra)
}
\description{
The Africa Infrastructure Country Diagnostic (AICD) was an unprecedented knowledge program
on Africa’s infrastructure that grew out of the pledge by the G8 Summit of 2005 at Gleneagles
to substantially increase ODA assistance to Africa, particularly to the infrastructure sector,
and the subsequent formation of the Infrastructure Consortium for Africa (ICA).
The AICD study was founded on the recognition that sub-Saharan Africa (SSA) suffers from
a very weak infrastructural base, and that this is a key factor in the SSA region failing
to realize its full potential for economic growth, international trade, and poverty reduction.
The study broke new ground, with primary data collection efforts covering network service
infrastructures (ICT, power, water & sanitation, road transport, rail transport, sea transport,
and air transport) from 2001 to 2006 in 24 selected African countries. Between them,
these countries account for 85 percent of the sub-Saharan Africa population, GDP,
and infrastructure inflows. The countries included in the initial study were: Benin,
Burkina Faso, Cameroon, Cape Verde, Chad, Côte d’Ivoire, Democratic Republic of Congo,
Ethiopia, Ghana, Kenya, Lesotho, Madagascar, Malawi, Mozambique, Namibia, Niger, Nigeria,
Rwanda, South Africa, Senegal, Sudan, Tanzania, Uganda, and Zambia. The study also
represents an unprecedented effort to collect detailed economic and technical data on African
infrastructure in relation to the fiscal costs of each of the sectors, future sector investment
needs, and sector performance indicators. As a result, it has been possible for the first time
to portray the magnitude of the continent’s infrastructure challenges and to provide detailed
and substantiated estimates on spending needs, funding gaps, and the potential efficiency dividends
 to be derived from policy reforms.
}
\examples{
data(infra)
}
\references{
African Development Bank Group
Published by source: 01 August 2019
(\href{http://infrastructureafrica.org/dataquery}{Original source})
}
\keyword{datasets}
